+++ src/digit_util.h
@@ -10,8 +10,8 @@ private:
 public:
   static const int typeVoltage = 1;
   static const int typeCurrent = 2;
-  void separate(int *v, int *mv, int *uv, bool *neg, float rawMv);
-  void separate2(int *v, int *mv, int *uv, int *nV, bool *neg, float rawMv);
+  void separate(int *v, int *mv, int *uv, bool *neg, double rawMv);
+  void separate2(int *v, int *mv, int *uv, int *nV, bool *neg, double rawMv);
   void renderValue(int x,int y,float val, int size, const int type);
   int blinkColor(int colorHigh, int colorLow, unsigned int period);
